+++ default.nix
@@ -2,5 +2,4 @@
 
 import ./lib {
   inherit (pkgs) lib newScope;
-  mkMyPkgs = callPackage: import ./pkgs callPackage;
 }