+++ lib/default.nix
@@ -1,5 +1,4 @@
 { lib
-, mkMyPkgs
 , newScope
 }:
 
@@ -7,7 +6,7 @@ lib.makeScope newScope (self:
 let
   inherit (self) callPackage;
 in
-(mkMyPkgs callPackage) // {
+{
   appendCrateRegistries = input: self.overrideScope' (final: prev: {
     crateRegistries = prev.crateRegistries // (lib.foldl (a: b: a // b) { } input);
   });
@@ -19,10 +18,13 @@ in
   cargoClippy = callPackage ./cargoClippy.nix { };
   cargoDoc = callPackage ./cargoDoc.nix { };
   cargoFmt = callPackage ./cargoFmt.nix { };
+  cargoHelperFunctionsHook = callPackage ./setupHooks/cargoHelperFunctions.nix { };
   cargoNextest = callPackage ./cargoNextest.nix { };
   cargoTarpaulin = callPackage ./cargoTarpaulin.nix { };
   cleanCargoSource = callPackage ./cleanCargoSource.nix { };
   cleanCargoToml = callPackage ./cleanCargoToml.nix { };
+  configureCargoCommonVarsHook = callPackage ./setupHooks/configureCargoCommonVars.nix { };
+  configureCargoVendoredDepsHook = callPackage ./setupHooks/configureCargoVendoredDeps.nix { };
   crateNameFromCargoToml = callPackage ./crateNameFromCargoToml.nix { };
 
   crateRegistries = self.registryFromDownloadUrl {
@@ -32,8 +34,11 @@ in
 
   downloadCargoPackage = callPackage ./downloadCargoPackage.nix { };
   downloadCargoPackageFromGit = callPackage ./downloadCargoPackageFromGit.nix { };
-  findCargoFiles = callPackage ./findCargoFiles.nix { };
   filterCargoSources = callPackage ./filterCargoSources.nix { };
+  findCargoFiles = callPackage ./findCargoFiles.nix { };
+  inheritCargoArtifactsHook = callPackage ./setupHooks/inheritCargoArtifacts.nix { };
+  installCargoArtifactsHook = callPackage ./setupHooks/installCargoArtifacts.nix { };
+  installFromCargoBuildLogHook = callPackage ./setupHooks/installFromCargoBuildLog.nix { };
   mkCargoDerivation = callPackage ./mkCargoDerivation.nix { };
   mkDummySrc = callPackage ./mkDummySrc.nix { };
 
@@ -46,6 +51,7 @@ in
 
   registryFromDownloadUrl = callPackage ./registryFromDownloadUrl.nix { };
   registryFromGitIndex = callPackage ./registryFromGitIndex.nix { };
+  removeReferencesToVendoredSourcesHook = callPackage ./setupHooks/removeReferencesToVendoredSources.nix { };
   urlForCargoPackage = callPackage ./urlForCargoPackage.nix { };
   vendorCargoDeps = callPackage ./vendorCargoDeps.nix { };
   vendorCargoRegistries = callPackage ./vendorCargoRegistries.nix { };